MB-59579 Disable assertShardMappings at lifecycle manager
If rebalance fails while shard token is in state ShardTokenRecoverShard
(before IsPendingReady is set to true), then rebalance would cleanup
the index on the destination. However, it is not possible to know
if all the indexes of a shard are deleted or not without additional
book-keeping at lifecycle manager. This makes clearing the book-keeping
a challenge.

Since the shard mappings are not cleared, they can be stale. During next
rebalance the stale book-keepings can cause false positives in assertions
i.e. shard-1 may be mapped to alt-1 in stale book-keeping but in current
rebalance, shard-1 can be mapped to alt-2 (E.g., due to replica repair).

Hence, disabling the assertions as no issues have been found in this area
so far (as adding additional book-keeping to maintain the book-keeping
correctly requires more code change which I believe is unnecessary at this
point).

Change-Id: I2e3dbe42166120399e9dd6707cdb4722691f3f39