MB-59460 Handle index drop during rebalance transfer/restore/recovery
A collection drop can happen during any during rebalance. When
changing RState of the instances, rebalancer should keep a track
of all the instances which are deleted as a result of collection
drop and then skip updating the RState of such instances.

a. If a collection drop happens during shard transfer, then source
node will skip the slice closure. Destination node will realize this
during recovery. Recovery would skip this index and shard rebalancer
adds this index to it's book-keeping. When changing RState of all the
indexes, this instance would be skipped by rebalancer

b. If a collection drop happens while restore is in progress, source
node would drop the instance and close the slices. Destination
behaviour would be the same as that of collection drop during transfer

c. If a collection drop happens after recovery is done (or) build is in
progress, then indexer would keep a track of the instances which are
dropped during rebalance. When RState change is triggered by rebalancer,
then indexer would check if the instance is dropped during rebalance and
if it is true, then indexer would return ErrIndexDeletedDuringRebal to
caller. Caller would handle this error

d. If rebalance fails after a token is moved to ShardTokenMerge (or) with
IsPendingReady = true, then the first cleanup after failure would change
the RState. Indexer will not clear droppedInsts book-keeping unless cleanup
is attempted once. During that time, RState change would check for dropped
instances and handles the ErrIndexDeletedDuringRebal error.

Change-Id: I9d006ba837fb14f7a21a6f50d036f93a22b00b9a