MB-59603 Wait for partition merge to finish before processing another instance
Otherwise, if partition merge is delayed because of flush in
progress, rebalancer would spawn multiple go-routines due to
the loop in waitForIndexState.

Also, cleanup book-keeping once instance is processed. Otherwise,
rebalancer will consider that the index is not processed and would
continue to wait for index to get processed.

Also, increase the logging frequency of "destTokenToMergeOrReady"
to 5 min from 20 seconds.

Change-Id: I7e8335bdc27469769e05926a519521ed239167e1