MB-59623 Always access storageMgrCmdCh in indexer main loop
Indexer code is built on a invariant that all message exchanges
with storageMgrCmdCh happen via indexer mainloop.

However, the "sendBucketNameNumVBucketsMap" method which access
storageMgrCmdCh is called inside "initPartnInst" which is spawned
in a go-routine for asynchronous recovery of indexes. This leads
to a violation of the invariant and in this ticket, it has been
observed that the LockShards command sent to storage manager
command channel has been consumed by this go-routine leading
to a rebalance hang.

This patch fixes the issue by calling "sendBucketNameNumVBucketsMap"
only in indexer mainloop even during async recovery.

Change-Id: Icbb3e9def4f328b9b1dcd67b784c7c0a0d377721