MB-59313: Fix index creation priority for non-serverless
The incoming index creation requests use the current time as request
timestamp. With this a potential starvation of scheduled indexes is avoided.
But there is a possibility of incoming request getting deprioritised if a
background index creation request with higher priority (older timestamp) is
triggered. When this happens in serverless, index creation gets scheduled in
the background. But in non-serverless, ability to background creation of
deprioritised indexes is not implemented. So, in non-serverless, ongoing
index creation fails.

In non-serverless, starvation of index creation requests is not a serious problem
as there is no multi-tenancy. So, using higher priority (zero timestamp) for ongoing
index creation request to avoid de-prioritisation.

Change-Id: I964b63c2dc64d987365943b041a51817aacbbda8