MB-59027 Control logging due to mergePartition
If a partition merge can not be served immediately due to an
on-going flush, then the partition merge will be attempted after
flush is done.

However, when there are multiple buckets, partition merge of one
bucket is attempted when flush for another bucket is done. This
will lead to multiple warning messages which can easily flood the
logs. This patch moves such warning messages to verbose messages
to control logging.

Change-Id: Id5a8dbe6aeb40518a5d001828c9308469799786f