MB-59655 Update in-memory book-keeping for dropOnSourceToken processing
When there is a dropOnSource state, source node is not updating its
book-keeping with the latest state. This results in source node having
state transfer token state.

If any error happens while processing dropOnSource state, then the token
would be updated with a wrong state and this can lead inconsistent
rebalance behaviour.

The fix for this issue is to update the in-memory book-keeping when
processing drop on source state.

Change-Id: Ic8d6c74a7ac8217d9abdc5e567ca2591f5b78389