MB-59622 Unlock all locked shards at the end of rebalance
In https://review.couchbase.org/c/indexing/+/184938, all
shards that have been locked on source have been unlocked at
the end of rebalance. This will help clear the lock for the
shards that have been locked if rebalance fails in restore/
recovery phase.

The change was applicable only for serverless deployments.
This patch makes it applicable for on-prem as well.

Also, made a minor change in shard rebalancer to update
instTransferPhase if either shard token is in restore (or)
recovery phase.

Change-Id: I7166bd909966dfeae54568350e10467fa4f91c43