MB-59717 Handle target inst. drop during rebalance
When an index is dropped during rebalance, it is tracked in
in indexer via "droppedIndexesDuringRebal" book-keeping.

When an recovery is triggered during rebalance, either realInstId
or proxyInstId can be picked up for recovery depending on the
presence of realInstId on the node. In case, realInstId is picked
up, indexer would just update the RState of the index without doing
any merge.

However, in the preValidateMergePartition phase, indexer only checks
for proxyInstId for merge. This means, if indexer recovers using
realInstId, then the proxy will not be part of "droppedIndexesDuringRebal"
and this can lead to merge failing and indexer crashing.

To handle such cases, indexer should also check for tgtInstId in
droppedIndexesDuringRebal map.

Change-Id: I5fd6f45a657d5ca1da0093899248923be62e056f