MB-59732 Do not drop replicas before grouping
Consider the following case: A shard pair has 10 indexes.
Of these 10 indexes, 5 indexes have replicas on all other
indexer nodes in the cluster. 5 indexes do not have replicas
on all nodes.

When this shard pair is moved as a part of rebalance, then
planner will drop the 5 replicas containing replicas on all
nodes and generate a token only for the 5 indexes that do not
have replicas. At the end of rebalance, the shards will be
dropped on the source. This will lead to a panic on indexer
because there are indexes belonging to the shard on the source
node which are not moved.

To avoid such cases, planner should not drop replicas before
index grouping happens. Droppinng replicas can happen after
index grouping.

Change-Id: I79348988a205fdadd609f90940badd54a3c4a013