MB-59675 Skip processing handleCleanupIndexFromInvalidKeyspace when rebalance is in progress
When rebalance is in progress, if "handleCleanupIndexFromInvalidKeyspace" is
triggered, then lifecycle manager would update the state of the indexes
in the keyspace to INDEX_STATE_DELETED but indexer would not process
the request (as this would be considered as a user initiated drop and
indexer would reject user initiated drop requests during rebalance).

This would lead to an inconsistency in metadata where lifecycle manager
marks the index as DELETED but the index would be alive and active from
indexer point of view. With shard rebalance, when the shard containing
such instances is moved, the indexes would be dropped on source. Drop
would be considered successful as the state of the index is DELETED in
metadata. The drop never reaches indexer for the same reason. So, when
the shard is destroyed, the corresponding index would be alive and
indexer would panic.

Change-Id: I23c10773dddc8caebbfe753ea7ddf6023219aed8