MB-59715 Skip indexes that are not moving in filterSolution
Consider the following case:

a. A shard proxy (replica: 0) is moving from n1 to n2
b. Replica 1 of same shard proxy exists on node n1 and it
   is not moving.

When filterSolution processes indexes, it should check cyclic,
unnecessary movements only for indexes in movement i.e. for replica 0.
Since that check does not happen explicitly, currently the check
happens for replica: 0 shard proxy as well. This will lead to
wrong token generation and rebalance would fail with "index already
exists" error.

To handle such issues, skip processing the index that is not moving
as a part of rebalance.
Change-Id: I580a41aa8454dc986c026455224b6beeb426a0ab

Change-Id: I7f8a9337dfbfdfe1b3c2805d8bd983809ba930b3