MB-55628: Fixed incorrect violation checks during Rebalance - Single Node removal - Placement of a Replica across SGs should not fail due to EquivIdx
Note: This patch addresses only 1 node removal cases, if the more than 1 node are getting removed the rebalance failures can still occur. In this case the customer has to drop replica instances or just trigger a rebalance with one node removal

Background:
The incorrect check happens when an Indexer Node is moving out i.e. we have non-zero OutIndexes. The rebalance failure is hit when for any one of the outIndex, the number of equivalent index is <= numLiveNode and there is atleast a ServerGroup with no replica of this index BUT the Server Group with no replica has Equivalent index on all the nodes in this SG.
This patch does not focus on index creation through planner but just the rebalance failures. Greedy Planner changes have been addressed in a separate patch already

What causes the rebalance failure?
1. As numEquivIndex <= numLiveNode we never make suppressEquivIndex Check to be true
2. When the Index is getting placed on a SG with Replica present, and as atleast one SG is available with No Replica -> Planner raises ServerGroupViolations or ReplicaViolation. Planner believes the SG with No Replica should be the recipient of this Index, maintaining HA across ServerGroups
3. When the index placement on a Node present in SG with no replica is tried, and if the Node has an equivalent Index -> Planner raises EquivalentIndex Violation as suppressEquivIndex is not set to true [as mentioned in pt 1.]
Due to the violations and no possible the placement of an Index from a deletedNode, the rebalance fails

Approach:
The Replicas are treated as first class citizens, meaning that a replica can be placed on a Node with equivalent Index for ensuring availability across Server Groups. Hence if we find a ServerGroup which does not have a replica but has Equivalent Index, we will ensure that this ServerGroup gets a replica, the movement of replica to the ServerGroup should not be withheld because of an Equivalent Index

Solution:
For an outgoing index, if the total number of nodes in the SGs-with-no-replica are less than the num of replicas present only on the deletedNodes, then suppress equivalent index check for this Index.

Change-Id: I868fa035566f1c53123071b6663861fc3611c9c1