MB-55628 Test for Equiv Index placement during Rebalance - Single Node Removal
Introduces test cases when a delete node has equivalent index with it's replica in the other SGs. These outIndexes need to be placed on the remaining nodes. The test cases are designed such that there are enough empty nodes to NOT suppress Equivalent Index checks, at the same time force the index placement in position where an outgoing index needs to be placed on an existing equivalent index.

Change-Id: I9aebb6aa4435a48f63a1dfea7115b9aaadfb9dd4