MB-58275 Check for pending drop cleanup before rebalance
Indexer handles the storage cleanup of index drop asynchronously. This
means that there is a small window of time during which index metadata
doesn't exist in the lifecycle manager and indexer book-keeping but the
actual storage clean has not been done yet.

If rebalance retry(e.g. after a failure), it is possible that the next
rebalance is successful in creating the index metadata while the pending
async cleanup clears the storage.

To avoid this situation, every rebalance should check if there is any
drop cleanup pending.

Change-Id: I394acd943bfde5b712c1e1f6ea779ecee4691cf2