MB-59814 Handle empty partitions during partition prune
The following sequence of events during rebalance can trigger
a panic loop on indexer:

a. Consider a partitioned index on 2 nodes: n1, n2
b. Add node n3 and trigger rebalance. Rebalancer will redistribute
   the partitions
c. Two shard tokens are generated for partitioned index movement
   n1 -> n3 (token-1) and n2 -> n3 (token-2). Indexer picks the instance
   in token-1 for realInst and builds the instance in token-2 as proxy
d. Once both the tokens move to ShardTokenMerge state, the destination
   node crashes
e. As a part of cleanup after restart, assume changing the RState of
   token-1 to fail (E.g., collection drop for the remaining instances
   on token-1 can lead to RState failure) but changing RState of
   token-2 does not fail
f. RState change of token-2 does not complete as RState change of
   token-1 did not succeed. As the instance from token-1 would not be in
   REBAL_ACTIVE, merge does not go through
g. As a part of cleanup of token-1, all partitions on token-1 would be
   pruned. The instance would not be deleted because there is a valid
   proxy from token-2

At this point, the realInst remains in REBAL_PENDING phase and merge
will never finish for the instance on token-2. This is part-1 of the fix.

Now, after (g), if indexer crashes, then there will be empty partitions
for an instance and indexer would end-up in panic loop in bootstrap. The
second part of the fix is to ignore such instances from further processing
and change the RState of such instance.

Change-Id: I7dbfa3f26033cf81aa00ea5003b781eba360b316