MB-59886 Use cloned version of stats for async recovery
During rebalance, index drop/cleanup after partition merge can happen
while async recovery is in progress. This will lead to a case where
the index stats map is being updated by drop/partition cleanup while
async recovery is reading the stats map. This can lead to a concurrent
map write and read error.

The fix for this issue is to use cloned version of the stats for async
recovery as async recovery only needs the pointer to the partition
stats of a particular instance which will not be chaned with other
accesses to the stats map.

Change-Id: I5d66c16a6c5bcbffa9f2e0d277bcf937cfadf5dd