MB-59677 Restore and unlock shards for tokens in Ready state
During cleanup in rebalance service manager, if the decision to
dropOnSource or dropOnDest has happened, then the shards can be
restored and unlocked on corresponding source or destination
without waiting for rebalance service manager to restore and
unlock shards.

Note: This is just a defensive fix as the change introduced in
https://review.couchbase.org/c/indexing/+/201143 will restore
and unlock shards anyways.

Change-Id: I1bc1e0935bb614fc3a9ed220918dd9b1785ddc23