MB-59875 Fail rebalance if cleanup is pending before acquiring lock
Currently, prepareRebalance() will wait for service manager mutex
and then fails a rebalance if cleanup is pending. But, it is possible
that cleanup can take time (Upto several minutes as observed in this
case) and this can lead to PrepareTopologyChange timing out. This
leads to incorrect message returned to user.

This patch addresses the rebalance failure message by checking for
cleanup pending without acquiring service manager mutex lock. If
there is an error, the caller of prepareRebalance() will set the
state as balanced = false. Hence, the update of this variable is
skipped if cleanup is found to be pending.

Change-Id: I94052a3f57a7833dc03065ab1ad4005b54df9a1b