MB-55160 add cancelRecoveryCh to plasma slice for shard rebalance
bg:
- recovery on shard can take long time; if rebalance is cancelled
during that time, then we don't have a way to stop this async recovery
once the async recovery is finished, index will dropped and cleanup
is finished; this makes cleanup a long running process.

- it can also happen that index gets dropped while async recovery is
going on in which recovery is of no use and will waste resources.

exp: API to cancel remains as follows:
- pass cancelCh from LifecycleMgr to plasma stores.
- on drop index if inst is in async recovery, trigger async cancel
by closing the passed channel.
- index delete happens once the async recovery is finished.

Change-Id: I6881136442785ef1ec2cdc15e4f48335518e0e48