MB-59945 Update transfer phase with instId on source node during repair
During replica repair, it is possible that the instanceID's in the
transfer token get renamed to suite the replica partitions on other
nodes in the cluster. Currently, all instances in transfer are being
tracked via InstIds, RealInstIds in the transfer token. If an index
drop were to happen (due to collection drop), then the closure of
index slice for instances that are in transfer phase is skipped.

However for replica repair, as instance ID's of source node need not
a part of the transfer token, rebalancer is not preventing the slice
closure. This is leading to rebalance failure with shard metadata
mismatch error.

This patch addresses this issue by populating the original instanceID's
of the instances on source node for "instsTransferPhase" book-keeping.

Change-Id: Ie6416495e3e9cdb2abc3cb63808c357b6184c7eb