MB-59715 Update instanceIds suiting replicas when assigning alternate shardIds
When assining slots to indexes, the replicaIds of the indexes will
be changed to suite the slot distribution on the node. When there
are partitioned indexes involved, this leads to a case where the
index instanceId and corresponding replicaId do not match i.e.
instance Id for replica: 0 can get slot belonging to replica: 1 and
vice versa.

This patch fixes the issue by changing the index instance Id along
with replicaId at the time of assigning alternate shardIds.

Change-Id: Ia8137610c68eb7458410b68c9f04ff11717b7183