MB-59280 Populate alternate shardIds for all indexes getting restored
Currently, if indexes are restored to a non-empty node, then the
alterante shardIds of those indexes will be reset. If such indexes
are not moved out of the node during planing phase, then alterante
shardIds for those indexes will not be populated.

This patch fixes the issue by populating alternate shardIds for
all the indexes that are getting restored during rebalance.

Change-Id: Ic5bbd8dee5b01baf697aaec3870dce0647f4d919