MB-59808 Do not clear acceptedNames book-keeping for scheduled index request
Currently, when a PREPARE request is made to lifecycle manager,
the "acceptedNames" map is updated with request definition ID.
This is done to prevent concurrent creations of same index. If
any other client tries to create same index (i.e. same keyspace
and same name), then it will be treated as duplicate and response
will be sent back to indexer.

If client knows that the index is not going to be scheduled, then
the acceptedNames book-keeping is cleared. For indexes getting
scheduled, this book-keeping should not be cleared. Due to a bug
in the code, the book-keeping is being cleared for scheduled indexes.
When another node tries to create index while the book-keeping is
cleared, during this duration, then the index originally scheduled
for creation will fail with duplicate index name error. In worst case,
both requests may not proceed forward (as seen in this ticket) when
there are multiple indexer nodes involved (as one indexer node can
accept one request and another indexer node can accept other request).

This patch addresses the issue by not clearing the book-keeping
of acceptedNames for scheduled indexes.

Change-Id: I3ed6128c732fbd5b4343055d98ade8c655164c6a