MB-59911 Destroy empty shards periodically
When all the instances on a shard are removed, the shard remains
on a disk. With frequent drops and creates, the number of shards
on a node will continue to raise with time. To address this issue,
indexer will periodically drop all the empty shards (every 30min).

Indexer will use plasma's API: GetCurrentEmptyShardInfo() to get
the list of empty shards. This API has to be called in indexer
main-loop and storage manager main-loop. Calling this API in
indexer main loop ensures that no new create operations can
happen while this API is being called (Otherwise, shard destroy
can race with index creation on same shard). Also, calling this
API in storage manager mainloop ensures that there won't be any
rollback happening on any of the shards.

The instance path on a shard is removed only when the instance
is completely destroyed (i.e. not just soft delete). So, async
slice.Close()/slice.Delete() from indexer will not race with
shard delete as plasma would serialize access to the shard.

Change-Id: I858ec5596a3dae1d7a28d8512933c5ba3edbbcd1