MB-60254 Update RState in metadata for an instance with zero partitions
The fix for MB-59814 does update the RState of an index with zero
partitions to RebalActive only in index instance map. If another
rebalance were to happen while the index is in this state, planner
would skip the index from further processing as the RState of the
index in metadata is still RebalPending. This will lead to rebalance
panic as the shards will be moved while an index exists on the shard
(since planner does not consider the index when planning for movement)

This patch fixes the issue by updating the RState in index metadata
as well along with the update in index instance map.

Change-Id: Icd5b6f7e232ba056993b83fc5966ecbcc2883071