MB-60257 Increase the size of bootstraps channel
In 5K test, after rebalance failure indexer ended up in a deadlock
due to the following sequence of events:

a. Lifecycle manager acquired meta_repo lock and is trying to
   update the topology (For index deleted that happened as a
   part of rebalance cleanup)

b. In the meantime, the bootstraps channel at lifecycle manager
   became full

c. Leader is trying to write a request to bootstraps channel but
   it is stuck as the channel is full

This has resulted in a deadlock. The fix for the issue is to
increase the size of bootstrap channel to 10000 from 1000

Change-Id: I4fe3d4d59c8a119425b99be6f74305f523734eb2