MB-60282 Return scope/collection not found error incase of mismatch in UID
The following sequence of events can lead to a rebalance failure:

a. Destination node has indexes on a collection: "c"
b. During rebalance, this collection was dropped
c. Before indexer got a chance to cleanup the indexes belonging to
   this collection, rebalance tried to recover the index
d. When indexer queries ns_server for this scope and collection,
   ns_server would return ID_NIL as the keyspace does not exist

However, as there are valid definitions for this collection, currently
indexer would return "Collection/Scope does not exist and temporarily..."
message. This would cause rebalance failure as rebalancer can only handle
ErrCollectionNotFound (Collection Not Found) (or) ErrScopeNotFound (Scope
Not Found) errors.

To address the rebalance failure, lifecycle manager would not check the
error and if the scope & collection ID's are nil and if the error message
says "Does not exist", it would return "Collection Not Found" error so
that rebalancer can handle the error effectively.

Change-Id: If62be9a65c5ccde67ce291820a553d187b8987ed