MB-59654 Handle snapEnd less than previously received snapEnd
Stream reader and timekeeper code are built on the assumption
that a new snapEnd is always greater than the previously
received snapEnd. However, this assumption is not always true.

It is possible for KV to send a snapEnd < prevSnapEnd on a stream
restart but the new snapEnd is always greater than the start
seqno requested. This patch addresses the issue by adjusting
the keyspaceIdFilter and keyspaceIdPrevSnapMap to the new
snapshot boundaries when such an snapshot is seen.
Timekeeper logic would also adjust its book-keeping when such
a snapshot is seen.

Change-Id: I1fe8e14cf0f1bb805c024be2dd589ffce047b658