MB-60336 Drain batchBuildReqCh if rebalance is cancelled
Following race condition exists in code:

a. Rebalance cancel is received and processBatchBuildReq method exits
b. From multiple tokens, startShardRecoveryNonServerless is trying to
   write to batchBuildReqCh
c. Since there is no consumer, the channel gets filled and eventually
   startShardRecoveryNonServerless will be blocked
d. This will lead to rebalance Cancel() getting stuck as cancel gets
   finished only after all go-routines gracefully exit. In this case,
   few go-routines are waiting to write to a channel

The fix for this issue is to drain the batchBuildReqCh even if the
"processBatchBuildReq" method exits. Once all go-routines exit,
the batchBuildReqCh will be closed

Change-Id: I828d64f1738d9ca975c432d11443d06717756a6a