MB-59256 Check for shutdown errors for every 1000 rows scanned
Currently, if majority of the scan is involved in filtering rows,
client can read timeout on the scan. However, server will continue
to process the scan as the check for error happens only when
server tries to write to client - which does not happen when
server is filtering all the rows.

To avoid such cases, server now checks if the scan has been shutdown
for every 1000 rows scanned. This will help server to fail fast
instead of scanning long ranges. The check has been made generic
so that it gets applied to long aggregation scans as well.

Change-Id: I133d9f9c1b50047f09dc251a48ad5f1aa6b5aa6a