MB-59654 Add isPersistanceActive REST endpoint
In perf tests for shard rebalance, it has been observed that
disk snapshotting is in progress at the start of rebalance. This
is leading to a case where the indexes are getting re-built on
destination node and not really capturing the performance
benefits for file based rebalance.

Hence, adding an isPersistanceActive endpoint so that perf team
can ensure that disk snapshot of an built index is available
by the time rebalance starts.

Change-Id: I574ca9dd4dc37365583d7d253ebd9fb3ce85e86b