MB-60534 Return cleanup in progress if there are any transfer tokens pending cleanup
Currently, the cleanupPending flag is reset if the global rebalance
token is removed by master inspite of transfer tokens being present
in metaKV. It is possible that the nodes are cleaning up the transfer
tokens due to a failed rebalance. If the cleanup races with a new
rebalance request, then the new rebalance request may timeout.

To avoid such spurious timeouts, QE team is using the "rebalanceCleanupStatus"
endpoint to check the cleanup status on each node and triggers a new
rebalance only if the cleanup is done on all nodes. So far the
status is updated only based on "cleanupPending" flag. From now on,
followers will check the presence of rebalance tokens in metaKV
and return the status accordingly.

Change-Id: Idb274ac4363457210cb51246e6be67ce3fe3b77a