MB-60561: Decrement stat tracking current len of TsQueue
Problem:
Currently when an element is removed from the TsList, the stat for
tracking current size of TsQueue is not decreased.

This means that the currInitTsQueueLen and currMaintTsQueueLen stat are
monotonically increasing variable. This can lead to unnecessary calls
to mergeTsQueue() even when the threshold for the max allowed size of
TsQueue has not reached.

Fix:
Decrement the currInitTsQueueLen or currMaintTsQueueLen when an element
is removed from TsList

Change-Id: I98b8669a52b9e96510d96d9534768b772d1e1729