MB-60575 Add all shard tokens to cleanup list before DCP tokens
In the current cleanup code, it is possible for a shard transfer
token and an empty node batch DCP token can come out of order.
If the empty node batch DCP token is dependant on a real instance
in shard token, then merge can remain stuck for ever.

As shard rebalancer always processes shard tokens first, cleanup
should also happen in the same order. This patch ensures that all
shard tokens are added to the head of the cleanup list which are
followed by empty node batch DCP tokens.

Change-Id: Icc7b4a5c2d4ec0fd2b5e8c8f9181157ef0b9d8e4