MB-60585 Populate indexDefn in instance for proxy usage
In some methods (E.g., dropReplicaIfNecessary), name of the indexUsage
is derived from Instance.Defn. For proxy, index defn is not populated.
This can lead to all shard proxies sharing the same name which is
incorrect.

This patch popualtes the name and replica paramenters inside the
Defn field of index instance of proxy usage. Also, the naming
convention of a shard proxy is fixed to make it similar to the
naming convention of other indexes.

Change-Id: I01a8f0cf3e2e299f33af45c262641270c3b1ea88