MB-60600: Call destTokenToMergeOrReady for all index insts in tt
If two transfer token have partitions of the same indexes, and if more
more than one such index exist in these transfer tokens, there can be a
circular dependencies where each index partition is waiting for the
partition in the other transfer token to be either merged or have its
RState changed to RebalActive. Causing a deadlock

fix:
Moving the wait group outside of the for loop, ensures that every index in the tt gets a chance for getting merged/RebalActive. This will prevent the deadlock

Change-Id: I9108a79622a8a5460e57838e297d5a712f3f908d