MB-60666 Initialize indexer and bucket stats for empty partition instances
Consider the following case for an index with 2 partitions

a. PartnId: 1 of an instance is moving from node A to node B

b. PartnId: 2 of the same instance is moving from node C to node A

c. If partnId: 1 movement finishes first, then the real instance
   will not be removed on node A as there is a proxy waiting for
   merge

If indexer crashes after (c), there will be an instance with empty
partition. Stats are not initialized for such partitions. If flush
happens, then since stats are not set, storage manager panics.

This patch initializes the indexer stats for such instances.

Change-Id: I912d4d9a747e56a2baed4e5302eb98c69a73485c