MB-60418 Preprocess unexploded secExprs and exprStatement
When there are many flattened array indexes involved, getIndexStatus
has to construct the index statement and unexploded secondary
expressions every time. This can lead to getIndexStatus timing out.

As index definition statement and the unexploded expressions do
not change once the index definition is created, this patch
pre-populates the Unexploded version of secondary expressions
and also the expression statement so that getIndexStatus need
not re-compute it every time.

Currently, only for indexes created in 7.6+ version will have the
pre-processed information.

Change-Id: I670ae91c98bd28015463c0d38a4c89b1bae43b7c