MB-60759 Skip updating RState for deferred indexes in ShardTokenMerge state
RState of deferred indexes are updated immediately after recovery.
Therefore, it is redundant to update RState again in ShardTokenMerge
state. If a collection were to be dropped before ShardTokenMerge,
indexer will not keep a track of deferred indexes (as RState is not REBAL_PENDING) and rebalancer will try to update the RState of index in this state. This will lead to index not being found (as indexer deleted the index due to collection drop) and a panic.

This patch fixes the issue by skipping RState change in ShardTokenMerge
state for deferred indexes.

Change-Id: I139607cff25d0abefbadabf6c11a5fd8fdef3dad