MB-60758 Prune slots which are on target nodes with different replicaId
Consider the following slot placement:

a. n0 (replica 0), n1 (replica 2), n3 (replica 3)
(Replica 1 is missing as node is failed over). Only
node n0 is full

b. Planner wants to place an index on n0 (repicaId: 0)
and n3 (replicaId: 1)

c. When pruning the slot based on global distribution,
the slot will not be pruned as the slot exists on n3
inspite of the slot having a different replicaID than
intended for.

d. This will lead to planner using wrong slot for the index
and wrong instance ID will be derived for the slot.

The patch fixes the issue by pruning such slots.

Change-Id: I5f5a9d2da8750dcf24a4939332560fbaa01f4fdb