MB-60756 filterSolution before populating alternate shardIds
When populating alternate shardIds, planner expects the slot
to shard affinity to remain constant. In the following case,
this can be violated:

a. There exists 3 nodes in the cluster: n0, n1, n2
b. n0 hosts all indexes with replicaId: 0, n1 hosts replicaId: 1,
   n2 hosts replicaId: 2. For this example, all these indexes
   share the same slot.
c. During rebalance, planner decided to move indexes belonging to
   this slot from n0 to n2 and n2 to n0 - resulting in a cyclic
   movement. This is valid according to planner. filterSolution
   will filter out such movements
d. After planning operation is done, following placement is decided
   by planner:
   n0 -> indexes with replicaId: 2, n1: indexes with replicaId: 1,
   n2 -> indexer with replicaId: 0
e. A new index with replicas: 0, 1 and 2 on nodes (n2, n1, n0) can
   pick this slot during rebalance

In this case, rebalance would fail with "index alredy exists" error.
This is because, the cyclic movement of shard proxies will be filtered
out in "filterSolution" method. Therefore, for the slot n0: hosts
replicaId: 0 as before. However, the new index has replicaId: 2
on node n0. So, rebalancer attempts to move the new index from n0 -> n2
and n2 -> n0 to match the alternate shardIds. But since the repliacs
already exists on these nodes, rebalance would fail.

It is incorrect for planner to select the slot with cyclic movements
in step (e) as this can lead to cyclic movements for new indexes
choosing the slots which can not be avoided as well (as alternate
shardIds have to be mapped). To avoid such cases, this patch fixes
the issue by filtering the solution from cyclic and redundant movements
before populating alternate shardIds.

Change-Id: I61cc9415bcb1788f3e3b1002aba89db9cc30203d