MB-60771 Reset alternateShardIds in instance defn during restore
Currently, when index meta is converted to index usage, the
index.Instance.Defn.AlternateShardIds map is shared across
all indexUsage structures (irrespective of the replica or
partition). ConvertToIndexUsage does *defn when assigning to
index.Instance.Defn which will create separate definitions
for all indexes but since *defn does a shallow copy,
AlternateShardIds will be shared

This is leading to a case where the alternate shardIds modified
by index usage with replicaId: 0 index are getting reflected in
index usage with replicaId: 1. The fix for this issue is to
re-initialize the AlternateShardIds in buildIndexHostMapping.

The first partition for any replica will always reset the map
and all further parititons of the same replica will reuse the map.

Change-Id: I0eb246d400654276596c1258b7d4794a33b3fd56