MB-60962 Use initialNode from solution in filterSolution
As a part of MB-60756, filterSolution() started moving indexes
to the destination node so that alternate shardIds can be
generated on filtered placement. However, when moving the indexes,
filterSolution used index.initialNode as final destination.

Using this object is incorrect as this object is created before
the solution is cloned in adjustInitialSolutionIfNecessary.
Therefore, even for same node, index.initialNode is different
from the node in final solution. Therefore, moving the index
to index.initialNode will add the index to the wrong object
and the index will disappear from final solution. This can
lead to rebalance failures.

This patch fixes the issue by using the correct initial node.

Change-Id: I05b5d643513d3a248701debbdefe4dbdef02437e