MB-60984 Use RefCount based book-keeping for lock shards in transfer manager
The following race condition can prevent an index from getting created
on a shard:

a. Rebalancer decided to perform more than one Copy i.e. it
   has decided to copy to 2 or more nodes
b. The shard will be locked 2 or more times
c. Rebalance failed after transfer is complete
d. This will lead to RestoreAndUnlockShards() getting called as a
   part of cleanup - UnlockShards() will not be called on source
   as rebalance failed.
e. RestoreAndUnlockShards() unlocks shard only once. The shard is
   locked twice in Step (b)

Any new indexes scheduled for creation on this shard will fail
as the shard is locked. This can lead to index creation and
rebalance failures.

Change-Id: I7ab33d77dbc016af05b5183ff15abe45b5d77e57