MB-60962 GroupIndexes before filterSolution() if new alternate shardIds are required
Group indexes before filtering as NeedsReplan will un-group the
indexes. The purpose of this filterSolution() method is to achieve
final index placement so that the alternate shardIds are populated
correctly. If the indexes are not grouped in to shard proxy, then
filterSolution() can do incorrect filtering.

For example, consider the distribution:
-- node n1: i1 (replica 0), i2 (replica 0), i3 (replica 0), i4 (replica 0)
-- node n2: i1 (replica 1), i2 (replica 1), i3 (replica 1), i4 (replica 1)
-- node n3: i2 (replica 2), i3 (replica 2), i4 (replica 2)
-- node n4: i3 (replica 3), i4 (replica 3)

The indexes i1, i2, i3, i4 share the same slot (Say s1). Consider a new
node (n5) being added to the cluster.

If planner decides to move shard_proxy_s1 (replica 2) from n3 to n2
and shard_proxy_s1 (replica 1) from n2 to n5, then filterSolution()
is ideally expected to move shard_proxy_s1 (replica 2) from n3 to n5
 - as the replica of a proxy is moving across nodes. However, due to
ungrouping of indexes, the movement of i2, i3, i4 will be filtered
but i1 (replica 1) can still move from n2 to n5 (as it is not jumping
across nodes). This can lead subsequent failures as the shard on n2
which hosts i1 will be moved to n5 and gets destroyed on n2.

Hence, always do filterSolution() on grouped indexes

Change-Id: I116bfcd2e13ce6b82e6e6e868ca4bce8a04212c3